[Fix & reafactor]:Refactor Folder structure & fixed path issues
