+++ src/App.js
@@ -1,8 +1,6 @@
 import "./App.css";
-import Footer from "./components/Footer";
-import Home from "./components/Home";
-import Header from "./components/Header";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import { Home, Header } from "./pages/index.js";
 
 function App() {
   return (
@@ -10,16 +8,7 @@ function App() {
       <Router>
         <Header />
         <Routes>
-          <Route
-            path="/"
-            element={
-              <div>
-                <Home />
-                <Footer />
-              </div>
-            }
-          ></Route>
-          <Route path="/products" ></Route>
+          <Route path="/" element={<Home />}></Route>
         </Routes>
       </Router>
     </div>
