+++ src/components/Header/Header.jsx
@@ -1,7 +1,8 @@
 import React from "react";
-import {Link} from "react-router-dom"
+import { Link } from "react-router-dom";
 
 export default function Header() {
+
   return (
     <div>
       <>
@@ -11,8 +12,7 @@ export default function Header() {
             <Link to="/" className="hm-pg-link">
               <h1 className="flex-row web-logo">
                 <span className="material-icons web-logo-icon">
-                  {" "}
-                  sports_esports{" "}
+                  sports_esports
                 </span>
                 GameKart
               </h1>
@@ -26,11 +26,11 @@ export default function Header() {
               className="search-bar"
             />
           </div>
-          <a href="./screens/LoginPage.html">
+          <Link to="/login">
             <button className="btn primary-btn">LOGIN</button>
-          </a>
+          </Link>
           <div className="flex-row header-right">
-            <a href="./screens/wishlist.html">
+            <Link to="/wishlist">
               <div className="relative-container">
                 <i className="material-icons header-icon">
                   favorite_border
@@ -39,15 +39,15 @@ export default function Header() {
                   </div>
                 </i>
               </div>
-            </a>
-            <a href="./screens/mycart.html" className="cart">
+            </Link>
+            <Link to="/cart" className="cart">
               <div className="flex-row relative-container">
                 <i className="material-icons header-icon"> shopping_cart </i>
                 <div className="badge top-right lrg red-clr flex-row center-it">
-                  2
+                 2
                 </div>
               </div>
-            </a>
+            </Link>
             <i className="material-icons header-icon"> logout </i>
           </div>
         </header>
