+++ src/pages/Home/Home.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import HomeContent from "../../components/HomeContent/HomeContent.jsx";
+import Footer from "../../components/Footer/Footer.jsx";
+
+export const Home = () => {
+  return (
+    <div>
+      <HomeContent />
+      <Footer />
+    </div>
+  );
+};
+
+export default Home;
