+++ src/pages/index.js
@@ -1,0 +1,4 @@
+import Home from "./Home/Home"
+import Header from "../components/Header/Header";
+
+export  {Home,Header}
